+++ src/Modules/Maintenance/MaintenanceGateway.php
@@ -2,9 +2,11 @@
 
 namespace Foodsharing\Modules\Maintenance;
 
+use Carbon\Carbon;
 use Foodsharing\Modules\Core\BaseGateway;
 use Foodsharing\Modules\Core\DBConstants\Basket\Status;
 use Foodsharing\Modules\Core\DBConstants\Foodsaver\SleepStatus;
+use Foodsharing\Modules\Core\DBConstants\Quiz\SessionStatus;
 
 class MaintenanceGateway extends BaseGateway
 {
@@ -138,8 +140,8 @@ class MaintenanceGateway extends BaseGateway
 
 					AND
 						b.id IN(:storeIds)', [
-						':storeIds' => implode(',', $storeIds)
-					]);
+					':storeIds' => implode(',', $storeIds)
+				]);
 			}
 		}
 
@@ -207,5 +209,26 @@ class MaintenanceGateway extends BaseGateway
 	public function unsetUserPhotos(array $foodsaverIds): void
 	{
 		$this->db->update('fs_foodsaver', ['photo' => ''], ['id' => $foodsaverIds]);
+	}
+
+	/**
+	 * Updates all quiz sessions that were finished or aborted more than two weeks ago by setting
+	 * questions and answers to null. After this the quiz results can not be seen anymore.
+	 *
+	 * @return int the number of updated entries
+	 */
+	public function updateFinishedQuizSessions(): int
+	{
+		return $this->db->update(
+			'fs_quiz_session',
+			[
+				'quiz_result' => null,
+				'quiz_questions' => null,
+			],
+			[
+				'status' => [SessionStatus::FAILED, SessionStatus::PASSED],
+				'time_end <' => Carbon::now()->subWeeks(2)->format('Y-m-d H:i:s'),
+			]
+		);
 	}
 }